phase3.6: kebab-case filenames/dirs, Db→Database rename, trim stale Phase 5 comment reference
