+++ server/middleware/error-handler.ts
