+++ server/routes/auth-routes.ts
