+++ server/routes/movie-routes.ts
