+++ server/test-support/database.ts
@@ -3,17 +3,17 @@ import { MongoMemoryServer } from 'mongodb-memory-server'
 
 let mongod: MongoMemoryServer
 
-export async function connectTestDb(): Promise<void> {
+export async function connectTestDatabase(): Promise<void> {
 	mongod = await MongoMemoryServer.create()
 	await mongoose.connect(mongod.getUri())
 }
 
-export async function disconnectTestDb(): Promise<void> {
+export async function disconnectTestDatabase(): Promise<void> {
 	await mongoose.disconnect()
 	await mongod.stop()
 }
 
-export async function clearTestDb(): Promise<void> {
+export async function clearTestDatabase(): Promise<void> {
 	const { collections } = mongoose.connection
 
 	await Promise.all(Object.values(collections).map(collection => collection.deleteMany({})))
