+++ server/tmdb-wrapper/index.ts
@@ -1,0 +1,2 @@
+export {default as TmdbMovie} from './movie-db-wrapper'
+
