--- server/tmdbWrapper/index.ts
@@ -1,2 +1,0 @@
-export {default as TmdbMovie} from './movieDBWrapper'
-
